fix(code-runtime): share runtime exec argv contract
